// apps/mobile/src/types/navigation.ts
export type AuthStackParamList = {
  Splash: undefined;
  Login: undefined;
  Register: undefined;
  ForgotPassword: undefined;
};

export type AppTabParamList = {
  DashboardTab: undefined;
  HistoryTab: undefined;
  AlertsTab: undefined;
  LimitsTab: undefined;
  ProfileTab: undefined;
};

export type HistoryStackParamList = {
  HistoryHome: undefined;
  ConsumptionDetails: undefined;
};